requirements.txt
click
i3ipc
pyyaml
